Integrating changes for ASCI aggregation work.
Moved some of the functions from ad_read_coll.c (which were shared with
the write code) into ad_aggregate.c.  All functions that perform
calculations for the purpose of aggregation are in there now.

Added cb_config_list.c, which handles the gathering and parsing of
processor names and the creation of the ranklist for I/O aggregation.

Added the "hints" structure off the ADIO_File structure.  This is filled
in by functions in ad_hints.c.