added support for dynamically loading xio drivers for 3 network types based on the presence of an evironment variable:
MPIG_SYSTEM_NET: specifies the prefered driver for system networks
MPIG_LAN_NET: specifies the prefered driver for LANS
MPIG_WAN_NET: specifies the prefered driver for WANs

when 2 nodes are determined to be on any given network the corresponding
enviroment variable is checked.  if the prefered drivers match for both
nodes that driver will be used.  If no matches are found the default driver
(tcp) will be used.