1. MAJOR BUG -- Collective operations seem to be performing an extra packing for derived datatypes. The first packing is for finding the actual length of the data that is being communicated, while the second is for the actual communication. The local processor representation anyway keeps track of the actual size; we don't need to recalculate it. So, we can get rid of one of the two packings.
2. MINOR BUG -- For small messages or for small communicator sizes,
the collective communication directly uses send/recv operations and
relies on its packing/unpacking. However, this forces each
intermediate node to unpack the data and repack it for
communication. Since we are anyway packing at the root node for large
messages and large communicator sizes, we can do the same for small
messages.